+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added fullscreen mode in location selection dialog
+- Added a dialog that displays the changed distance and sequence after optimizing the itinerary
 
 ### Changed
 - Changed edit duration modal to use a more ui coherent timer picker